Profile Fragment Updates
Edit Profile will show the logged in name and phone number of user in the firebase. Also, modification to the signup to take also the name and phone number. Verify Dialog was created to secure the security of the user when changing password. Changing Password and Email is not working yet.